+++ src/Engine/Snapshots/EntitySnapshots.cpp
@@ -1650,11 +1650,11 @@ void reconstruct(const PlayerFrame_MM7 &src, PlayerFrame *dst) {
 void reconstruct(const LevelDecoration_MM7 &src, LevelDecoration *dst) {
     dst->uDecorationDescID = src.uDecorationDescID;
     dst->uFlags = LevelDecorationFlags(src.uFlags);
-    dst->vPosition = src.vPosition;
+    dst->vPosition = src.vPosition.toFloat();
     dst->_yawAngle = (TrigLUT.uIntegerHalfPi * src.field_1A) / 90;
     // src.field_1A - actually yaw angle in degrees, used when _yawAngle is not set.
     if (src._yawAngle)
-    dst->_yawAngle = src._yawAngle;
+        dst->_yawAngle = src._yawAngle;
     dst->uCog = src.uCog;
     dst->uEventID = src.uEventID;
     dst->uTriggerRange = src.uTriggerRange;
